fix: 로그인 후 새로고침 시 로그인 상태가 풀리는 버그 수정 및 리팩토링 (#43)
