+++ 003 Code/FE/notion-linked-blog/src/apis/apiClient.js
@@ -4,4 +4,5 @@ const {SERVER} = process.env;
 
 export const apiClient = axios.create({
 	baseURL: `${SERVER}`,
+	withCredentials: true,
 });
